Revert "Initial commit"
This reverts commit 6379e37dd5e84efc335fc2893ea2a21c26cfb160.
