--- .env
@@ -1,1 +1,0 @@
-PUBLIC_URL=.
