+++ package.json
@@ -7,13 +7,10 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "cheerio": "^1.0.0",
     "html2canvas": "^1.4.1",
-    "node-fetch": "^3.3.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-scripts": "5.0.1",
-    "react-toastify": "^11.0.2",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
