--- public/service-worker.js
@@ -1,17 +1,0 @@
-self.addEventListener('install', event => {
-    console.log('Service worker installing...');
-    self.skipWaiting();
-  });
-  
-  self.addEventListener('activate', event => {
-    console.log('Service worker activating...');
-  });
-  
-  self.addEventListener('fetch', event => {
-    event.respondWith(
-      caches.match(event.request).then(cachedResponse => {
-        return cachedResponse || fetch(event.request);
-      })
-    );
-  });
-
