+++ test_server/index.js
@@ -74,17 +74,6 @@ const transformOCRText = (text) => {
     return text.replace(/ /g, '').replace(/\//g, '7').replace(/!/g, '1').replace(/\|/g, '1');
 };
 
-const validateOCRText = async (text, imageBuffer) => {
-    if (text.length > 3) {
-        console.log('OCR result is invalid, retrying...');
-        const newText = await performOCR(imageBuffer);
-        if (newText) {
-            return transformOCRText(newText);
-        }
-    }
-    return text;
-};
-
 app.get('/images', async (req, res) => {
     const targetUrl = 'https://www.hanwha701.com';
     let page;
@@ -144,7 +133,7 @@ app.get('/images', async (req, res) => {
 
             let ocrResultText = await performOCR(croppedImageBuffer);
             if (ocrResultText) {
-                ocrResultText = await validateOCRText(ocrResultText, croppedImageBuffer);
+                ocrResultText = transformOCRText(ocrResultText);
             }
 
             const timestamp = new Date().toLocaleString();
