+++ test_server/package.json
@@ -12,9 +12,6 @@
   "dependencies": {
     "cors": "^2.8.5",
     "express": "^4.21.2",
-    "jimp": "^1.6.0",
-    "puppeteer": "^23.11.1",
-    "sharp": "^0.33.5",
-    "tesseract.js": "^5.1.1"
+    "puppeteer": "^23.11.1"
   }
 }
